Improved customizability of animations Improved documentation Internal refactoring and reorganization
