+++ lib/flutter_map.dart
@@ -25,8 +25,10 @@ export 'package:flutter_map/src/geo/latlng_bounds.dart';
 export 'package:flutter_map/src/gestures/interactive_flag.dart';
 export 'package:flutter_map/src/gestures/map_events.dart';
 export 'package:flutter_map/src/gestures/multi_finger_gesture.dart';
-export 'package:flutter_map/src/layer/attribution_layer.dart'
-    hide SourceAttribution;
+export 'package:flutter_map/src/layer/attribution_layer/animation.dart';
+export 'package:flutter_map/src/layer/attribution_layer/rich.dart';
+export 'package:flutter_map/src/layer/attribution_layer/simple.dart';
+export 'package:flutter_map/src/layer/attribution_layer/source.dart';
 export 'package:flutter_map/src/layer/circle_layer.dart';
 export 'package:flutter_map/src/layer/marker_layer.dart';
 export 'package:flutter_map/src/layer/overlay_image_layer.dart';
