+++ lib/src/layer/attribution_layer/rich.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/plugin_api.dart';
 import 'package:meta/meta.dart';
@@ -27,176 +26,32 @@ enum AttributionAlignment {
   final Alignment real;
 }
 
-/// A simple, classic style, attribution widget, to be placed in
-/// [FlutterMap.nonRotatedChildren]
+/// A prebuilt dynamic attribution layer that supports both logos and text
+/// through [SourceAttribution]s
 ///
-/// Displayed as a padded translucent [backgroundColor] box with the following
-/// text: 'flutter_map | © [source]', where [source] is wrapped with [onTap].
-class SimpleAttributionWidget extends StatelessWidget {
-  /// Attribution text, such as 'OpenStreetMap contributors'
-  final Text source;
-
-  /// Callback called when [source] is tapped/clicked
-  final VoidCallback? onTap;
-
-  /// Color of the box containing the [source] text
-  final Color? backgroundColor;
-
-  /// Anchor the widget in a position of the map
-  final Alignment alignment;
-
-  /// A simple, classic style, attribution widget, to be placed in
-  /// [FlutterMap.nonRotatedChildren]
-  ///
-  /// Displayed as a padded translucent white box with the following text:
-  /// 'flutter_map | © [source]'.
-  const SimpleAttributionWidget({
-    Key? key,
-    required this.source,
-    this.onTap,
-    this.backgroundColor,
-    this.alignment = Alignment.bottomRight,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => Align(
-        alignment: alignment,
-        child: ColoredBox(
-          color: backgroundColor ?? Theme.of(context).colorScheme.background,
-          child: GestureDetector(
-            onTap: onTap,
-            child: Padding(
-              padding: const EdgeInsets.all(3),
-              child: Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  const Text('flutter_map | © '),
-                  MouseRegion(
-                    cursor: onTap == null
-                        ? MouseCursor.defer
-                        : SystemMouseCursors.click,
-                    child: source,
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      );
-}
-
-/// Base class for attributions that render themselves as widgets
+/// [TextSourceAttribution]s are shown in a popup box that can be visible or
+/// invisible. Its state is toggled by a tri-state [openButton]/[closeButton] :
+///   1. Not hovered, not opened: faded button, invisible box
+///   2. Hovered, not opened: full opacity button, invisible box
+///   3. Opened: full opacity button, visible box
 ///
-/// Extended/implemented by [TextSourceAttribution] & [LogoSourceAttribution].
-@internal
-abstract class SourceAttribution extends StatelessWidget {
-  final VoidCallback? _onTap;
-
-  const SourceAttribution._({
-    super.key,
-    VoidCallback? onTap,
-  }) : _onTap = onTap;
-
-  Widget _render(BuildContext context);
-
-  @override
-  @nonVirtual
-  Widget build(BuildContext context) {
-    if (_onTap == null) return _render(context);
-
-    return GestureDetector(
-      onTap: _onTap,
-      child: MouseRegion(
-        cursor: SystemMouseCursors.click,
-        child: _render(context),
-      ),
-    );
-  }
-}
-
-/// A simple text attribution displayed in the popup box of a
-/// [RichAttributionWidget]
-class TextSourceAttribution extends SourceAttribution {
-  /// Default style used to display the [text], only when
-  /// [SourceAttribution._onTap] is not `null`
-  static const defaultLinkTextStyle =
-      TextStyle(decoration: TextDecoration.underline);
-
-  /// The text to display as attribution, styled with [textStyle]
-  final String text;
-
-  /// Style used to display the [text]
-  final TextStyle? textStyle;
-
-  /// Whether to add the '©' character to the start of [text] automatically
-  final bool prependCopyright;
-
-  /// A simple text attribution displayed in the popup box of a
-  /// [RichAttributionWidget]
-  const TextSourceAttribution(
-    this.text, {
-    super.key,
-    super.onTap,
-    this.prependCopyright = true,
-    TextStyle? textStyle,
-  })  : textStyle = textStyle ?? (onTap == null ? null : defaultLinkTextStyle),
-        super._();
-
-  @override
-  Widget _render(BuildContext context) => Text(
-        '${prependCopyright ? '© ' : ''}$text',
-        style: textStyle,
-      );
-}
-
-/// An image attribution permanently displayed adjacent to the open/close icon of
-/// a [RichAttributionWidget]
-class LogoSourceAttribution extends SourceAttribution {
-  /// The logo to display as attribution, usually an [Image.asset] or
-  /// [Image.network]
-  final Widget image;
-
-  /// Optional text to display inside a [Tooltip] when the logo is hovered over,
-  /// to provide extra clarity
-  final String? tooltip;
-
-  /// Height of the [image] (fitted into a [SizedBox])
-  ///
-  /// Should be the same as [RichAttributionWidget.permanentHeight], otherwise
-  /// layout issues may occur.
-  final double height;
-
-  /// An image attribution permanently displayed adjacent to the open/close icon
-  /// of a [RichAttributionWidget]
-  const LogoSourceAttribution(
-    this.image, {
-    super.key,
-    super.onTap,
-    this.tooltip,
-    this.height = 24,
-  }) : super._();
-
-  @override
-  Widget _render(BuildContext context) {
-    final sizedImage = SizedBox(height: height, child: image);
-    if (tooltip == null) return sizedImage;
-    return Tooltip(message: tooltip, child: sizedImage);
-  }
-}
-
-/// A prebuilt attribution layer that supports both logos and text through
-/// [SourceAttribution]s
+/// The hover state on mobile devices is unspecified, but the behaviour is
+/// usually inconsequential on mobile devices anyway, due to the fingertip
+/// covering the entire button.
 ///
-/// [TextSourceAttribution]s are shown in a popup box (toggled by a tap/click on
-/// the [openButton]/[closeButton]), unlike [LogoSourceAttribution], which are
-/// visible permanently adjacent to the open/close button.
+/// [LogoSourceAttribution]s are shown adjacent to the open/close button, to
+/// comply with some stricter tile server requirements (such as Mapbox). These
+/// are usually supplemented with a [TextSourceAttribution].
 ///
 /// The popup box also closes automatically on any interaction with the map.
 ///
-/// Shows a 'flutter_map' attribution logo and text, unless
-/// [showFlutterMapAttribution] is `false`.
+/// Animations are built in by default, and configured/handled through
+/// [RichAttributionWidgetAnimation] - see that class and the [animationConfig]
+/// property for more information. By default, a simple fade/opacity animation
+/// is provided by [FadeRAWA]. [ScaleRAWA] is also available.
 ///
-/// Can be further customized to a certain extent through the other properties.
+/// Read the documentation on the individual properties for more information and
+/// customizability.
 class RichAttributionWidget extends StatefulWidget {
   /// List of attributions to display
   ///
@@ -234,13 +89,13 @@ class RichAttributionWidget extends StatefulWidget {
   /// Whether to add an additional attribution logo and text for 'flutter_map'
   final bool showFlutterMapAttribution;
 
-  /// The curve to use when toggling the visibility of the popup box and the
-  /// state of the open/close icons
-  final Curve animationCurve;
-
-  /// The length of the animation that toggles the visibility of the popup box
-  /// and the state of the open/close icons
-  final Duration animationDuration;
+  /// Animation configuration, through the properties and handler/builder
+  /// defined by a [RichAttributionWidgetAnimation] implementation
+  ///
+  /// Can be extensivley customized by implementing a custom
+  /// [RichAttributionWidgetAnimation], or the prebuilt [FadeRAWA] and
+  /// [ScaleRAWA] animations can be used with limited customization.
+  final RichAttributionWidgetAnimation animationConfig;
 
   /// If not [Duration.zero] (default), the popup box will be open by default and
   /// hidden this long after the map is initialised
@@ -250,19 +105,32 @@ class RichAttributionWidget extends StatefulWidget {
   /// [LogoSourceAttribution].
   final Duration popupInitialDisplayDuration;
 
-  /// A prebuilt attribution layer that supports both logos and text through
-  /// [SourceAttribution]s
+  /// A prebuilt dynamic attribution layer that supports both logos and text
+  /// through [SourceAttribution]s
   ///
-  /// [TextSourceAttribution]s are shown in a popup box (toggled by a tap/click
-  /// on the [openButton]/[closeButton]), unlike [LogoSourceAttribution]s, which
-  /// are visible permanently in a row adjacent to the open/close button.
+  /// [TextSourceAttribution]s are shown in a popup box that can be visible or
+  /// invisible. Its state is toggled by a tri-state [openButton]/[closeButton] :
+  ///   1. Not hovered, not opened: faded button, invisible box
+  ///   2. Hovered, not opened: full opacity button, invisible box
+  ///   3. Opened: full opacity button, visible box
+  ///
+  /// The hover state on mobile devices is unspecified, but the behaviour is
+  /// usually inconsequential on mobile devices anyway, due to the fingertip
+  /// covering the entire button.
+  ///
+  /// [LogoSourceAttribution]s are shown adjacent to the open/close button, to
+  /// comply with some stricter tile server requirements (such as Mapbox). These
+  /// are usually supplemented with a [TextSourceAttribution].
   ///
   /// The popup box also closes automatically on any interaction with the map.
   ///
-  /// Shows a 'flutter_map' attribution logo and text, unless
-  /// [showFlutterMapAttribution] is `false`.
+  /// Animations are built in by default, and configured/handled through
+  /// [RichAttributionWidgetAnimation] - see that class and the [animationConfig]
+  /// property for more information. By default, a simple fade/opacity animation
+  /// is provided by [FadeRAWA]. [ScaleRAWA] is also available.
   ///
-  /// Can be further customized to a certain extent through the other properties.
+  /// Read the documentation on the individual properties for more information
+  /// and customizability.
   const RichAttributionWidget({
     super.key,
     required this.attributions,
@@ -273,8 +141,7 @@ class RichAttributionWidget extends StatefulWidget {
     this.popupBorderRadius,
     this.permanentHeight = 24,
     this.showFlutterMapAttribution = true,
-    this.animationCurve = Curves.easeInOut,
-    this.animationDuration = const Duration(milliseconds: 150),
+    this.animationConfig = const FadeRAWA(),
     this.popupInitialDisplayDuration = Duration.zero,
   });
 
@@ -323,10 +190,10 @@ class RichAttributionWidgetState extends State<RichAttributionWidget> {
   @override
   Widget build(BuildContext context) {
     final persistentAttributionItems = [
-      ...widget.attributions
-          .whereType<LogoSourceAttribution>()
-          .cast<Widget>()
-          .separate(SizedBox(width: widget.permanentHeight / 1.5)),
+      ...List<Widget>.from(
+        widget.attributions.whereType<LogoSourceAttribution>(),
+        growable: false,
+      ).interleave(SizedBox(width: widget.permanentHeight / 1.5)),
       if (widget.showFlutterMapAttribution)
         LogoSourceAttribution(
           Image.asset(
@@ -338,9 +205,9 @@ class RichAttributionWidgetState extends State<RichAttributionWidget> {
         ),
       SizedBox(width: widget.permanentHeight * 0.1),
       AnimatedSwitcher(
-        switchInCurve: widget.animationCurve,
-        switchOutCurve: widget.animationCurve,
-        duration: widget.animationDuration,
+        switchInCurve: widget.animationConfig.buttonCurve,
+        switchOutCurve: widget.animationConfig.buttonCurve,
+        duration: widget.animationConfig.buttonDuration,
         child: popupExpanded
             ? (widget.closeButton ??
                 (context, close) => IconButton(
@@ -387,12 +254,12 @@ class RichAttributionWidgetState extends State<RichAttributionWidget> {
           alignment: widget.alignment.real,
           children: [
             if (persistentAttributionSize != null)
-              AnimatedOpacity(
-                opacity: popupExpanded ? 1 : 0,
-                curve: widget.animationCurve,
-                duration: widget.animationDuration,
-                child: Padding(
-                  padding: const EdgeInsets.all(6),
+              Padding(
+                padding: const EdgeInsets.all(6),
+                child: widget.animationConfig.popupAnimationBuilder(
+                  context: context,
+                  isExpanded: popupExpanded,
+                  config: widget,
                   child: Container(
                     decoration: BoxDecoration(
                       color: widget.popupBackgroundColor ??
@@ -444,8 +311,8 @@ class RichAttributionWidgetState extends State<RichAttributionWidget> {
               cursor: SystemMouseCursors.click,
               child: AnimatedOpacity(
                 opacity: persistentHovered || popupExpanded ? 1 : 0.5,
-                curve: widget.animationCurve,
-                duration: widget.animationDuration,
+                curve: widget.animationConfig.buttonCurve,
+                duration: widget.animationConfig.buttonDuration,
                 child: Padding(
                   padding: const EdgeInsets.all(4),
                   child: FittedBox(
@@ -467,10 +334,10 @@ class RichAttributionWidgetState extends State<RichAttributionWidget> {
   }
 }
 
-extension _ListExt<E> on Iterable<E> {
-  Iterable<E> separate(E separator) sync* {
+extension _ListExt<E> on List<E> {
+  Iterable<E> interleave(E separator) sync* {
     for (int i = 0; i < length; i++) {
-      yield elementAt(i);
+      yield this[i];
       if (i < length) yield separator;
     }
   }
